requirements.txt
aiohttp==3.4.4
aiosqlite==0.8.0
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.0a1621+g66e5033
idna==2.7
idna-ssl==1.1.0
multidict==4.5.1
pycparser==2.19
PyNaCl==1.2.1
requests==2.20.1
six==1.11.0
urllib3==1.24.1
websockets==6.0
yarl==1.2.6
